Much faster implementation of groupBy
This approach involves creating a string representation
of the row then grouping on that rather taking
successive intersections of indices.

The approach has a few downsides:
* For very large values the row representation can become and unwieldy memory bottleneck.
* Allocating and concatenating that many strings isn't very memory efficient.

Run with profiling however this implementation reduces allocations from 73GB to ~1GB
on the examples in the data/folder.

The speed also reduces from ~5 minutes to ~20 seconds.