Parse directly to types.
Before we'd keep the intermediate strings and only
parse at the end. That turned out to be really
wasteful especially in the early part of the
program. This change drastically reduces the
memory footprint of the program but also
introduces a lot of untested logic. This
marks the end of the performance optimization
journey for now. On to testing.