feat: Better error messages from propagated exceptions. (#95)
Before we would throw in library code. This PR tries to reduce that and bubbles all exceptions in internal functions to user facing functions so error messaging can be improved.

The expression code is ugly and complicated but it's bad to look into the kitchen of your favourite restaurant anyway.